Put auto hazard pay in a separate file
